refactor: move app shell out of package init
Move the KleinanzeigenBot class from __init__.py (545 lines) into a
dedicated app.py module, leaving __init__.py as a thin 27-line package
initializer that re-exports the class.

- New app.py (525 lines): full KleinanzeigenBot class with all imports,
  constructor, run() command dispatch, load_ads(), and delegator methods
- __init__.py: reduced to colorama init, LOG setup, re-export, main()
  delegator, and __name__ guard
- Translations: split __init__.py section header, moved class-owned
  translations to new app.py section
- Tests: updated patch targets from package-root to concrete module paths
